ai-interview-backend-fastapi/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
httpx
